requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
customtkinter==5.2.2
cx_Freeze==7.2.3
cx_Logging==3.2.0
darkdetect==0.8.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
lief==0.15.1
mock==5.1.0
multidict==6.1.0
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
propcache==0.2.0
psutil==6.0.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pypiwin32==223
pyserial==3.5
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dotenv==1.0.1
pyusb==1.2.1
pywin32==307
pywin32-ctypes==0.2.3
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
shiboken6==6.7.2
urllib3==2.2.3
usb==0.0.83.dev0
wheel==0.44.0
WMI==1.5.1
yarl==1.17.1
